fix: landing page suspense boundary and header menu logout ux
